requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly.express
scikit-learn
numpy
joblib